Move u64 helpers to circuit_frame
