+++ src/circuit/gadgets/constraints.rs
@@ -1,8 +1,5 @@
 // Initially taken from: rust-fil-proofs/storage-proofs-core/src/gadgets/
 
-use crate::circuit::add_clause_single;
-use crate::circuit::gadgets::case::multi_case_aux;
-use crate::circuit::gadgets::case::CaseClause;
 use crate::circuit::gadgets::data::GlobalAllocations;
 use crate::circuit::gadgets::pointer::AllocatedPtr;
 use crate::field::LurkField;
@@ -18,37 +15,6 @@ use bellperson::{
     ConstraintSystem, SynthesisError,
 };
 use ff::PrimeField;
-use num_bigint::BigUint;
-use num_integer::Integer;
-use num_traits::FromPrimitive;
-
-#[derive(Default)]
-struct CompResults<'a, F: LurkField> {
-    same_sign: Vec<CaseClause<'a, F>>,
-    a_neg_and_b_not_neg: Vec<CaseClause<'a, F>>,
-    a_not_neg_and_b_neg: Vec<CaseClause<'a, F>>,
-}
-impl<'a, F: LurkField> CompResults<'a, F> {
-    fn add_clauses_comp(
-        &mut self,
-        key: F,
-        result_same_sign: &'a AllocatedNum<F>,
-        result_a_neg_and_b_not_neg: &'a AllocatedNum<F>,
-        result_a_not_neg_and_b_neg: &'a AllocatedNum<F>,
-    ) {
-        add_clause_single(&mut self.same_sign, key, result_same_sign);
-        add_clause_single(
-            &mut self.a_neg_and_b_not_neg,
-            key,
-            result_a_neg_and_b_not_neg,
-        );
-        add_clause_single(
-            &mut self.a_not_neg_and_b_neg,
-            key,
-            result_a_not_neg_and_b_neg,
-        );
-    }
-}
 
 /// Adds a constraint to CS, enforcing an equality relationship between the allocated numbers a and b.
 ///
@@ -502,153 +468,6 @@ pub fn must_be_simple_bit(x: &Boolean) -> AllocatedBit {
     }
 }
 
-// This function helps to enforce a  comparison relation between a and b.
-// It receives as input argument `diff`, which must be constrained to be
-// equal to the difference (a - b).
-// The last argument is `op2`, which can be <, <=, >, >=
-pub fn comparison_helper<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    g: &GlobalAllocations<F>,
-    a: &AllocatedNum<F>,
-    b: &AllocatedNum<F>,
-    diff: &AllocatedNum<F>,
-    op2: &AllocatedNum<F>,
-) -> Result<(Boolean, AllocatedPtr<F>, Boolean), SynthesisError> {
-    let a_is_negative = allocate_is_negative(&mut cs.namespace(|| "enforce a is negative"), a)?;
-    let b_is_negative = allocate_is_negative(&mut cs.namespace(|| "enforce b is negative"), b)?;
-    let diff_is_negative =
-        allocate_is_negative(&mut cs.namespace(|| "enforce diff is negative"), diff)?;
-
-    let diff_is_zero = alloc_is_zero(&mut cs.namespace(|| "diff is zero"), diff)?;
-
-    let diff_is_not_positive = or(
-        &mut cs.namespace(|| "diff is not positive"),
-        &diff_is_negative,
-        &diff_is_zero,
-    )?;
-
-    let diff_is_positive = diff_is_not_positive.not();
-
-    let diff_is_not_negative = diff_is_negative.not();
-
-    let not_one_negative_and_other_not_negative = Boolean::xor(
-        &mut cs.namespace(|| "not one negative and other not negative"),
-        &a_is_negative,
-        &b_is_negative,
-    )?
-    .not();
-
-    let a_negative_and_b_not_negative = Boolean::and(
-        &mut cs.namespace(|| "a negative and b not negative"),
-        &a_is_negative,
-        &b_is_negative.not(),
-    )?;
-
-    let alloc_num_diff_is_negative = boolean_to_num(
-        &mut cs.namespace(|| "Allocate num for diff_is_negative"),
-        &diff_is_negative,
-    )?;
-
-    let alloc_num_diff_is_not_positive = boolean_to_num(
-        &mut cs.namespace(|| "Allocate num for diff_is_not_positive"),
-        &diff_is_not_positive,
-    )?;
-
-    let alloc_num_diff_is_positive = boolean_to_num(
-        &mut cs.namespace(|| "Allocate num for diff_is_positive"),
-        &diff_is_positive,
-    )?;
-
-    let alloc_num_diff_is_not_negative = boolean_to_num(
-        &mut cs.namespace(|| "Allocate num for diff_is_not_negative"),
-        &diff_is_not_negative,
-    )?;
-
-    let mut comp_results = CompResults::default();
-    comp_results.add_clauses_comp(
-        Op2::Less.as_field(),
-        &alloc_num_diff_is_negative,
-        &g.true_num,
-        &g.false_num,
-    );
-    comp_results.add_clauses_comp(
-        Op2::LessEqual.as_field(),
-        &alloc_num_diff_is_not_positive,
-        &g.true_num,
-        &g.false_num,
-    );
-    comp_results.add_clauses_comp(
-        Op2::Greater.as_field(),
-        &alloc_num_diff_is_positive,
-        &g.false_num,
-        &g.true_num,
-    );
-    comp_results.add_clauses_comp(
-        Op2::GreaterEqual.as_field(),
-        &alloc_num_diff_is_not_negative,
-        &g.false_num,
-        &g.true_num,
-    );
-
-    let comparison_defaults = [&g.default_num, &g.default_num, &g.default_num];
-
-    let comp_clauses = [
-        &comp_results.same_sign[..],
-        &comp_results.a_neg_and_b_not_neg[..],
-        &comp_results.a_not_neg_and_b_neg[..],
-    ];
-
-    let comparison_result = multi_case_aux(
-        &mut cs.namespace(|| "comparison multicase results"),
-        op2,
-        &comp_clauses,
-        &comparison_defaults,
-        g,
-    )?;
-
-    let comp_val_same_sign_num = comparison_result.0[0].clone();
-    let comp_val_a_neg_and_b_not_neg_num = comparison_result.0[1].clone();
-    let comp_val_a_not_neg_and_b_neg_num = comparison_result.0[2].clone();
-    let is_comparison_tag = comparison_result.1.not();
-
-    let comp_val1 = pick(
-        &mut cs.namespace(|| "comp_val1"),
-        &a_negative_and_b_not_negative,
-        &comp_val_a_neg_and_b_not_neg_num,
-        &comp_val_a_not_neg_and_b_neg_num,
-    )?;
-    let comp_val2 = pick(
-        &mut cs.namespace(|| "comp_val2"),
-        &not_one_negative_and_other_not_negative,
-        &comp_val_same_sign_num,
-        &comp_val1,
-    )?;
-
-    let comp_val_is_zero = alloc_is_zero(&mut cs.namespace(|| "comp_val_is_zero"), &comp_val2)?;
-
-    let comp_val = AllocatedPtr::pick(
-        &mut cs.namespace(|| "comp_val"),
-        &comp_val_is_zero,
-        &g.nil_ptr,
-        &g.t_ptr,
-    )?;
-
-    Ok((is_comparison_tag, comp_val, diff_is_negative))
-}
-
-// Enforce 0 <= num < 2ˆn.
-pub fn enforce_at_most_n_bits<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    g: &GlobalAllocations<F>,
-    num: AllocatedNum<F>,
-    n: usize,
-) -> Result<(), SynthesisError> {
-    let num_bits = num.to_bits_le(&mut cs.namespace(|| "u64 remainder bit decomp"))?;
-    let v = num_bits[n..255].to_vec();
-    boolean_summation(&mut cs.namespace(|| "add all MSBs"), &v, &g.false_num);
-    Ok(())
-}
-
 // Enforce div and mod operation for U64. We need to show that
 // arg1 = q * arg2 + r, such that 0 <= r < arg2.
 pub fn enforce_u64_div_mod<F: LurkField, CS: ConstraintSystem<F>>(
@@ -791,147 +610,6 @@ pub fn allocate_is_negative<F: LurkField, CS: ConstraintSystem<F>>(
     Ok(num_is_negative.clone())
 }
 
-// ATTENTION:
-// Convert from bn to num. This allocation is NOT constrained here.
-// In the circuit we use it to prove u64 decomposition, since using bn
-// we have division with remainder, which is used to find the quotient
-// after dividing by 2ˆ64. Therefore we constrain this relation afterwards.
-pub fn allocate_unconstrained_bignum<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    bn: BigUint,
-) -> Result<AllocatedNum<F>, SynthesisError> {
-    let bytes_le = bn.to_bytes_le();
-    let mut bytes_padded = [0u8; 32];
-    bytes_padded[0..bytes_le.len()].copy_from_slice(&bytes_le);
-    let ff = F::from_bytes(&bytes_padded).unwrap();
-    let num = AllocatedNum::alloc(&mut cs.namespace(|| "num"), || Ok(ff)).unwrap();
-    Ok(num)
-}
-
-// Convert from num to u64.
-pub fn u64_op<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    g: &GlobalAllocations<F>,
-    maybe_u64: &AllocatedPtr<F>,
-    store: &Store<F>,
-) -> Result<AllocatedPtr<F>, SynthesisError> {
-    let p64_bn = BigUint::pow(&BigUint::from_u32(2).unwrap(), 64);
-    let v = match maybe_u64.hash().get_value() {
-        Some(v) => v,
-        None => F::zero(),
-    };
-    let u64_ptr = store.get_u64(v.to_u64_unchecked());
-    let field_bn = BigUint::from_bytes_le(v.to_repr().as_ref());
-
-    let (q_bn, r_bn) = field_bn.div_rem(&p64_bn);
-
-    let q_num = allocate_unconstrained_bignum(&mut cs.namespace(|| "q"), q_bn)?;
-    let r_num = allocate_unconstrained_bignum(&mut cs.namespace(|| "r"), r_bn)?;
-
-    // a = b.q + r
-    let product = mul(
-        &mut cs.namespace(|| "product(q,pow(2,64))"),
-        &q_num,
-        &g.power2_64_num,
-    )?;
-    let sum = add(&mut cs.namespace(|| "sum remainder"), &product, &r_num)?;
-    let u64_decomp = alloc_equal(
-        &mut cs.namespace(|| "check u64 decomposition"),
-        &sum,
-        maybe_u64.hash(),
-    )?;
-    enforce_true(
-        &mut cs.namespace(|| "enforce u64 decomposition"),
-        &u64_decomp,
-    )?;
-
-    // enforce remainder range
-    enforce_at_most_n_bits(
-        &mut cs.namespace(|| "remainder u64 range"),
-        g,
-        r_num.clone(),
-        64,
-    )?;
-
-    let output = AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "u64_op"), store, || Ok(&u64_ptr))?;
-    let u64_is_remainder = alloc_equal(
-        &mut cs.namespace(|| "check u64 value is the remainder"),
-        &r_num,
-        output.hash(),
-    )?;
-    enforce_true(
-        &mut cs.namespace(|| "enforce u64 decomposition output"),
-        &u64_is_remainder,
-    )?;
-    Ok(output)
-}
-
-pub fn hide<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    g: &GlobalAllocations<F>,
-    secret: &AllocatedNum<F>,
-    maybe_payload: &AllocatedPtr<F>,
-    store: &Store<F>,
-) -> Result<AllocatedPtr<F>, SynthesisError> {
-    AllocatedPtr::construct_commitment(
-        &mut cs.namespace(|| "commitment hash"),
-        g,
-        store,
-        secret,
-        maybe_payload,
-    )
-}
-
-pub fn num<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    maybe_num: &AllocatedPtr<F>,
-    store: &Store<F>,
-) -> Result<AllocatedPtr<F>, SynthesisError> {
-    let num_ptr = if let Some(ptr) = maybe_num.ptr(store).as_ref() {
-        let scalar_ptr = store.get_expr_hash(ptr).expect("expr hash missing");
-        match store.get_num(crate::Num::Scalar::<F>(*scalar_ptr.value())) {
-            Some(n) => n,
-            None => store.get_nil(),
-        }
-    } else {
-        store.get_nil()
-    };
-    AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "num"), store, || Ok(&num_ptr))
-}
-
-pub fn comm<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    maybe_comm: &AllocatedPtr<F>,
-    store: &Store<F>,
-) -> Result<AllocatedPtr<F>, SynthesisError> {
-    let hash = match maybe_comm.hash().get_value() {
-        Some(h) => h,
-        None => F::zero(), // dummy value
-    };
-    let comm_ptr = match store.get_maybe_opaque(Tag::Comm, hash) {
-        Some(c) => c,
-        None => store.get_nil(),
-    };
-    AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "open"), store, || Ok(&comm_ptr))
-}
-
-pub fn char_op<F: LurkField, CS: ConstraintSystem<F>>(
-    mut cs: CS,
-    maybe_char: &AllocatedPtr<F>,
-    store: &Store<F>,
-) -> Result<AllocatedPtr<F>, SynthesisError> {
-    let char_ptr = if let Some(ptr) = maybe_char.ptr(store).as_ref() {
-        let scalar_ptr = store.get_expr_hash(ptr).expect("expr hash missing");
-        match scalar_ptr.value().to_u32() {
-            Some(n) => store.get_char(char::from_u32(n).unwrap()),
-            None => store.get_nil(),
-        }
-    } else {
-        store.get_nil()
-    };
-    AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "char_op"), store, || Ok(&char_ptr))
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1061,48 +739,6 @@ mod tests {
     }
 
     #[test]
-    fn test_enforce_n_bits() {
-        let mut cs = TestConstraintSystem::<Fr>::new();
-        let s = &mut Store::<Fr>::default();
-
-        let g = GlobalAllocations::new(&mut cs.namespace(|| "global_allocations"), s).unwrap();
-
-        let num = crate::Num::<Fr>::Scalar(Fr::from_u64(42).unwrap()); // 42 = 101010
-        let alloc_num =
-            AllocatedNum::alloc(&mut cs.namespace(|| "num"), || Ok(num.into_scalar())).unwrap();
-
-        let res = enforce_at_most_n_bits(
-            &mut cs.namespace(|| "enforce at most n bits"),
-            &g,
-            alloc_num,
-            6,
-        );
-        assert!(res.is_ok());
-        assert!(cs.is_satisfied());
-    }
-
-    #[test]
-    fn test_enforce_n_bits_negative() {
-        let mut cs = TestConstraintSystem::<Fr>::new();
-        let s = &mut Store::<Fr>::default();
-
-        let g = GlobalAllocations::new(&mut cs.namespace(|| "global_allocations"), s).unwrap();
-
-        let num = crate::Num::<Fr>::Scalar(Fr::from_u64(42).unwrap()); // 42 = 101010
-        let alloc_num =
-            AllocatedNum::alloc(&mut cs.namespace(|| "num"), || Ok(num.into_scalar())).unwrap();
-
-        let res = enforce_at_most_n_bits(
-            &mut cs.namespace(|| "enforce at most n bits"),
-            &g,
-            alloc_num,
-            5,
-        );
-        assert!(res.is_ok());
-        assert!(!cs.is_satisfied());
-    }
-
-    #[test]
     fn test_enforce_u64_div_mod() {
         let mut cs = TestConstraintSystem::<Fr>::new();
         let s = &mut Store::<Fr>::default();
@@ -1156,60 +792,6 @@ mod tests {
         assert!(cs.is_satisfied());
         assert_eq!(q.hash().get_value(), Fr::from_u64(0));
         assert_eq!(r.hash().get_value(), Fr::from_u64(0));
-    }
-
-    #[test]
-    fn test_u64_op() {
-        let mut cs = TestConstraintSystem::<Fr>::new();
-        let s = &mut Store::<Fr>::default();
-
-        let g = GlobalAllocations::new(&mut cs.namespace(|| "global_allocations"), s).unwrap();
-
-        let a = s.num(42);
-        let alloc_a = AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "a"), s, || Ok(&a)).unwrap();
-
-        let a_u64 = u64_op(&mut cs.namespace(|| "u64 op"), &g, &alloc_a, s).unwrap();
-        assert!(cs.is_satisfied());
-        assert_eq!(a_u64.hash().get_value(), Fr::from_u64(42));
-    }
-
-    #[test]
-    fn test_u64_op_coercion() {
-        let mut cs = TestConstraintSystem::<Fr>::new();
-        let s = &mut Store::<Fr>::default();
-
-        let g = GlobalAllocations::new(&mut cs.namespace(|| "global_allocations"), s).unwrap();
-        let alloc_pow2_64 = AllocatedPtr::from_parts(&g.num_tag, &g.power2_64_num);
-
-        let a_u64 = u64_op(&mut cs.namespace(|| "u64 op"), &g, &alloc_pow2_64, s).unwrap();
-        assert!(cs.is_satisfied());
-        assert_eq!(a_u64.hash().get_value(), Fr::from_u64(0));
-    }
-
-    #[test]
-    fn test_enforce_comparison() {
-        let mut cs = TestConstraintSystem::<Fr>::new();
-        let s = &mut Store::<Fr>::default();
-
-        let g = GlobalAllocations::new(&mut cs.namespace(|| "global_allocations"), s).unwrap();
-
-        let a = s.num(42);
-        let alloc_a = AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "a"), s, || Ok(&a)).unwrap();
-        let b = s.num(43);
-        let alloc_b = AllocatedPtr::alloc_ptr(&mut cs.namespace(|| "b"), s, || Ok(&b)).unwrap();
-        let diff = sub(&mut cs.namespace(|| "sub"), alloc_a.hash(), alloc_b.hash()).unwrap();
-
-        let (_, comp_val, _) = comparison_helper(
-            &mut cs.namespace(|| "enforce u64 div mod"),
-            &g,
-            &alloc_a.hash(),
-            &alloc_b.hash(),
-            &diff,
-            &g.op2_less_tag,
-        )
-        .unwrap();
-        assert!(cs.is_satisfied());
-        assert_eq!(comp_val.hash().get_value(), g.t_ptr.hash().get_value());
     }
 
     #[test]
